Moved to using Boost to reduce complexity.
There were at least two large parts of the program that I was doing manually that boost does better, so it is a good time to switch.

I also moved Spine out of it's separate thread (it now runs directly in the main thread), removed some superfluous objects that we don't need and changed the socket types for module message functions to use an enum.