Changed module compilation logic to introduce author namespaces.
This is purely a filesystem organisation change, to make it cleaner to split up mainline modules from user-produced modules. It should also allow re-use of module names (liek 'config') since they'll be bound to their author/group (e.g. libmainline_config).

The location of the modules on disk is now a compile-time option of the Spine. Rather than relying on the 'modules' being below the current directory (that also contains the binary), we now specify the location in the CMakeLists file and compile it in. This should lead to good defaults being changed based on make-type -- 'make' will be different to 'make install'. This will eventually be the same for all program locations, binary/logs/modules/config.